backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
hvac==1.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.1
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
